Lab example activities cleanup to fit enhanced SymbolLayer example
