+++ MapboxAndroidDemo/src/gpservices/AndroidManifest.xml
@@ -401,14 +401,5 @@
             </intent-filter>
         </activity>
 
-        <activity
-            android:name=".labs.MapillaryActivity"
-            android:label="@string/activity_labs_mapillary_title"
-            android:screenOrientation="portrait">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.mapbox.mapboxandroiddemo.MainActivity"/>
-        </activity>
-
     </application>
 </manifest>
