+++ MapboxAndroidDemo/src/main/AndroidManifest.xml
@@ -171,13 +171,6 @@
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity" />
         </activity>
         <activity
-            android:name=".labs.RecyclerViewSymbolLayerActivity"
-            android:label="@string/activity_lab_symbol_layer_on_map_title">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.mapbox.mapboxandroiddemo.MainActivity" />
-        </activity>
-        <activity
             android:name=".labs.AnimatedImageSourceActivity"
             android:label="@string/activity_labs_gif_on_map_title">
             <meta-data
@@ -545,7 +538,7 @@
         <activity
             android:name=".account.LoadingActivity"
             android:screenOrientation="portrait" />
-        <activity android:name=".labs.ProjectPinpointSymbolActivity"
+        <activity android:name=".labs.SymbolLayerExampleActivity"
             android:screenOrientation="portrait"/>
     </application>
