--- MapboxAndroidDemo/src/main/res/layout/activity_lab_mapillary.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:mapbox="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="16dp"
-        android:layout_marginRight="16dp"
-        android:layout_marginEnd="16dp"
-        android:src="@drawable/ic_crop_square_white_24dp"
-        mapbox:fabSize="normal"
-        mapbox:layout_constraintBottom_toBottomOf="parent"
-        mapbox:layout_constraintRight_toRightOf="parent"/>
-
-    <com.mapbox.mapboxsdk.maps.MapView
-        android:id="@+id/mapView"
-        mapbox:mapbox_cameraTargetLat="55.6089295863"
-        mapbox:mapbox_cameraTargetLng="13.0006076843"
-        mapbox:mapbox_cameraZoom="11"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-
-</android.support.constraint.ConstraintLayout>
