+++ MapboxAndroidDemo/src/main/res/layout/activity_symbol_layer_example.xml
