+++ MapboxAndroidDemo/src/main/res/values/descriptions_strings.xml
@@ -75,7 +75,7 @@
     <string name="activity_lab_los_angeles_tourism_description">Use the style API to highlight parks, hotels, and attractions. A pulsing animation is also added to the colors.</string>
     <string name="activity_lab_indoor_map_description">Display an indoor map of a building with toggles to switch between floor levels.</string>
     <string name="activity_lab_rv_on_map_description">Manipulate the map based on recyclerview interactions.</string>
-    <string name="activity_lab_symbol_layer_on_map_description">Add markers via SymbolLayer and manipulate the map based on recyclerview interactions.</string>
+    <string name="activity_lab_symbol_layer_on_map_description">Add markers via SymbolLayer and manipulate the data in real time.</string>
     <string name="activity_lab_mapillary_description">Add Mapillary vector tiles to a map.</string>
     <string name="activity_labs_inset_map_description">Show a smaller inset map fragment and link it to a larger map for two map interaction. Great for gaming!</string>
     <string name="activity_labs_gif_on_map_description">Show an animated image (GIF) anywhere on the map</string>
