+++ MapboxAndroidDemo/src/main/res/values/titles_strings.xml
@@ -66,15 +66,14 @@
     <string name="activity_image_generator_snapshot_notification_title">Snapshot Notification</string>
     <string name="activity_image_generator_snapshot_share_title">Share snapshot image</string>
     <string name="activity_lab_location_picker_title">Location picker</string>
-    <string name="activity_lab_rv_on_map_title">Recyclerview interaction</string>
-    <string name="activity_lab_symbol_layer_on_map_title">Recyclerview + SymbolLayer markers</string>
+    <string name="activity_lab_rv_on_map_title">RecyclerView interaction</string>
+    <string name="activity_lab_symbol_layer_on_map_title">SymbolLayer example + Mapillary integration</string>
     <string name="activity_lab_marker_following_route_title">Marker following route</string>
     <string name="activity_lab_space_station_location_title">Space station current location</string>
     <string name="activity_lab_picture_in_picture_title">Picture in picture</string>
     <string name="activity_lab_off_route_title">Off route detection</string>
     <string name="activity_lab_los_angeles_tourism_title">Los Angeles for tourists</string>
     <string name="activity_lab_indoor_map_title">Indoor Map</string>
-    <string name="activity_labs_mapillary_title">Mapillary integration</string>
     <string name="activity_labs_inset_map_title">Inset map</string>
     <string name="activity_labs_gif_on_map_title">Animated image source (GIF)</string>
